refactor: pacakge 이동
- factory, service 클래스를 domain 패지키로 이동
