+++ src/main/java/lotto/domain/WinningBallFactory.java
@@ -1,4 +1,4 @@
-package lotto.factory;
+package lotto.domain;
 
 import java.util.Arrays;
 import java.util.List;
