+++ src/test/java/lotto/domain/LottoResultsTest.java
@@ -6,7 +6,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import lotto.dto.WinningInfo;
-import lotto.factory.WinningBallFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
